Completely removed drw.
Since we now support EMWH for struts (kinda) we can get rid of
the built in panel and use an application built for that
purpose. This is more unixy.

I use `gnome-panel`, but I'm sure there is a more minimal EWMH
supporting panel for the minimalists. Heck, we could rewrite
the dwm panel as a standalone.